Time: 2 ms (55.05%), Space: 42.8 MB (91.44%) - LeetHub

+++ 1945-sum-of-digits-of-string-after-convert/1945-sum-of-digits-of-string-after-convert.java
@@ -2,7 +2,7 @@ class Solution {
     public int getLucky(String s, int k) {
         StringBuilder sb = new StringBuilder();
         for(int i = 0; i < s.length(); i++){
-            sb.append((int)s.charAt(i)-'a'+1);
+            sb.append(s.charAt(i)-'a'+1);
         }
         String s1 = sb.toString();
         int sum=0;
